+++ README.md
@@ -1,6 +1,6 @@
-# homework1
-otus.spring.homework1
+# homework
+otus.spring.homework
 
 
 1) Запуск через jar доступен если собрать приложение при помощи maven-assembly-plugin, командой mvn package
-2) Запуск через идею осущещсввляется через метод main() в ru.otus.spring.homework1.Main
+2) Запуск через идею осущещсввляется через метод main() в ru.otus.spring.homework.Main
