+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ru.khorunzhev</groupId>
-    <artifactId>otus.homework1</artifactId>
+    <artifactId>otus.homework</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -61,11 +61,11 @@
                             <goal>single</goal>
                         </goals>
                         <configuration>
-                            <finalName>otus_spring_khorunzhev_homework1</finalName>
+                            <finalName>otus_spring_khorunzhev_homework</finalName>
                             <archive>
                                 <manifest>
                                     <mainClass>
-                                        ru.otus.spring.homework1.Main
+                                        ru.otus.spring.homework.Main
                                     </mainClass>
                                 </manifest>
                             </archive>
