+++ src/main/java/ru/otus/spring/homework/model/FreeFormQuestion.java
@@ -1,0 +1,12 @@
+package ru.otus.spring.homework.model;
+
+import lombok.Getter;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+public class FreeFormQuestion extends Question {
+
+    private final String answer;
+
+}
