+++ src/main/java/ru/otus/spring/homework/model/Survey.java
@@ -1,0 +1,12 @@
+package ru.otus.spring.homework.model;
+
+import lombok.Data;
+import lombok.experimental.SuperBuilder;
+
+import java.util.List;
+
+@Data
+@SuperBuilder
+public class Survey {
+    private final List<Question> questions;
+}
