+++ src/main/java/ru/otus/spring/homework/model/enums/SurveyCSVHeaders.java
@@ -1,0 +1,8 @@
+package ru.otus.spring.homework.model.enums;
+
+public enum SurveyCSVHeaders {
+    Question,
+    QuestionType,
+    NumberOfAnswers,
+    CorrectAnswer;
+}
