+++ src/main/java/ru/otus/spring/homework/serivce/answers/AnswerCheckService.java
@@ -1,0 +1,6 @@
+package ru.otus.spring.homework.serivce.answers;
+
+public interface AnswerCheckService {
+
+    boolean checkAnswer(String expectedAnswer, String actualAnswer);
+}
