+++ src/main/java/ru/otus/spring/homework/serivce/answers/AnswerCheckServiceImpl.java
@@ -1,0 +1,12 @@
+package ru.otus.spring.homework.serivce.answers;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AnswerCheckServiceImpl implements AnswerCheckService {
+
+    @Override
+    public boolean checkAnswer(String expectedAnswer, String actualAnswer) {
+        return expectedAnswer.equals(actualAnswer);
+    }
+}
