+++ src/main/java/ru/otus/spring/homework/serivce/questions/AskQuestionService.java
@@ -1,0 +1,11 @@
+package ru.otus.spring.homework.serivce.questions;
+
+import ru.otus.spring.homework.model.Question;
+
+import java.util.List;
+
+public interface AskQuestionService {
+
+    void askAllQuestions(List<Question> questionList);
+    String askQuestion(Question question);
+}
