+++ src/main/java/ru/otus/spring/homework/serivce/survey/SurveyCheckService.java
@@ -1,0 +1,7 @@
+package ru.otus.spring.homework.serivce.survey;
+
+public interface SurveyCheckService {
+
+    void runSurvey();
+
+}
