+++ src/main/java/ru/otus/spring/homework/serivce/utils/InteractWithUserService.java
@@ -1,0 +1,7 @@
+package ru.otus.spring.homework.serivce.utils;
+
+public interface InteractWithUserService {
+
+    void writeTo(String info);
+    String readFrom();
+}
