+++ src/main/java/ru/otus/spring/homework/serivce/utils/exceptions/ReadUserInformationException.java
@@ -1,0 +1,7 @@
+package ru.otus.spring.homework.serivce.utils.exceptions;
+
+public class ReadUserInformationException extends RuntimeException {
+    public ReadUserInformationException(Exception e) {
+        super(e);
+    }
+}
