--- src/main/java/ru/otus/spring/homework1/dao/QuestionDao.java
@@ -1,11 +1,0 @@
-package ru.otus.spring.homework1.dao;
-
-import ru.otus.spring.homework1.model.Question;
-
-import java.util.List;
-
-public interface QuestionDao {
-
-    List<Question> findAll();
-
-}
