--- src/main/java/ru/otus/spring/homework1/model/FreeFormQuestion.java
@@ -1,14 +1,0 @@
-package ru.otus.spring.homework1.model;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@SuperBuilder
-public class FreeFormQuestion extends Question {
-
-    private final String answer;
-
-}
