--- src/main/java/ru/otus/spring/homework1/model/enums/SurveyCSVHeaders.java
@@ -1,8 +1,0 @@
-package ru.otus.spring.homework1.model.enums;
-
-public enum SurveyCSVHeaders {
-    Question,
-    QuestionType,
-    NumberOfAnswers,
-    CorrectAnswer;
-}
