--- src/main/java/ru/otus/spring/homework1/serivce/answers/AnswerCheckService.java
@@ -1,6 +1,0 @@
-package ru.otus.spring.homework1.serivce.answers;
-
-public interface AnswerCheckService {
-
-    boolean checkAnswer(String expectedAnswer, String actualAnswer);
-}
