--- src/main/java/ru/otus/spring/homework1/serivce/answers/AnswerCheckServiceImpl.java
@@ -1,12 +1,0 @@
-package ru.otus.spring.homework1.serivce.answers;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class AnswerCheckServiceImpl implements AnswerCheckService {
-
-    @Override
-    public boolean checkAnswer(String expectedAnswer, String actualAnswer) {
-        return expectedAnswer.equals(actualAnswer);
-    }
-}
