--- src/main/java/ru/otus/spring/homework1/serivce/questions/AskQuestionService.java
@@ -1,13 +1,0 @@
-package ru.otus.spring.homework1.serivce.questions;
-
-import lombok.extern.java.Log;
-import ru.otus.spring.homework1.dao.QuestionDao;
-import ru.otus.spring.homework1.model.Question;
-
-import java.util.List;
-
-public interface AskQuestionService {
-
-    void askAllQuestions(List<Question> questionList);
-    String askQuestion(Question question);
-}
