--- src/main/java/ru/otus/spring/homework1/serivce/survey/SurveyCheckService.java
@@ -1,7 +1,0 @@
-package ru.otus.spring.homework1.serivce.survey;
-
-public interface SurveyCheckService {
-
-    void runSurvey();
-
-}
