--- src/main/java/ru/otus/spring/homework1/serivce/utils/InteractWithUserService.java
@@ -1,7 +1,0 @@
-package ru.otus.spring.homework1.serivce.utils;
-
-public interface InteractWithUserService {
-
-    void writeTo(String info);
-    String readFrom();
-}
