--- src/main/java/ru/otus/spring/homework1/serivce/utils/exceptions/ReadUserInformationException.java
@@ -1,7 +1,0 @@
-package ru.otus.spring.homework1.serivce.utils.exceptions;
-
-public class ReadUserInformationException extends RuntimeException {
-    public ReadUserInformationException(Exception e) {
-        super(e);
-    }
-}
